+++ assets/javascripts/kitten/components/cards/donation-card.test.js
@@ -25,38 +25,9 @@ describe('<DonationCard />', () => {
 
     const component = renderer
       .create(
-        <DonationCard
-          titleAmount="Custom title mount"
-          textDescription="Custom text description"
-        />,
+        <DonationCard title="Custom title mount" button="Custom text button" />,
       )
       .toJSON()
-
-    it('matches with snapshot', () => {
-      expect(component).toMatchSnapshot()
-    })
-  })
-
-  describe('with some props', () => {
-    beforeEach(() => {
-      component = renderer
-        .create(
-          <DonationCard
-            titleAmount="Custom title mount"
-            imageProps={{
-              src: '#image',
-              alt: 'Image alt',
-            }}
-            textDescription="Custom text description"
-            button="Custom text button"
-            myContribution="Custom my contribution"
-            manageContribution="Custom manage contribution"
-            manageContributionLink="Custom manage link contribution"
-            renderAmountInput={() => <div>Custom function</div>}
-          />,
-        )
-        .toJSON()
-    })
 
     it('matches with snapshot', () => {
       expect(component).toMatchSnapshot()